Set up Ticket Contract Set up Ticket Test

+++ contracts/Migrations.sol
@@ -1,9 +1,9 @@
-// SPDX-License-Identifier: GPL-2.0-or-later
+// SPDX-License-Identifier: MIT
 pragma solidity >=0.4.22 <0.9.0;
 
 contract Migrations {
   address public owner = msg.sender;
-  uint public last_completed_migration;
+  uint256 public last_completed_migration;
 
   modifier restricted() {
     require(
